Lots of general cleanups and fixes, haven't looked at the code in a while
The \samepage LaTeX directive appears to work well spacing
problems so that the workspace is not broken on a page boundary.

The Makefile for the diffeqs example problems was removing
the .tex files after each build.  It seems to have something
to do with the auto generated rules (%).  Changing them to
an explicit form fixed the issue.

Refined the formatting of the LaTeX output.
Added Vim folds so the source is easier to navigate.
Also had to add a pattern substitution to fix the use of }}}
in math equations which conflict with Vim folds.