fixed test header title truncation at edge of page
Previously the test header was formatted using constant spacing.
And if a long title was used it would be truncated at the edge
of the page.

This has been fixed so that the header is the full width of the
page and the size will be automatically calculated.  This was
accomplished by using the 'tabularx' package.