refactored using Python, YAML and Mustache
Instead of using a home brew [config] format,
YAML is now used for the test problems and
.test files.  This allows for blocks of LaTeX
to be easily used without escaping.

The test.template's now uses Mustache so the
problems, answers, titles are easily plugged
in using the test config.  This also makes it
easier to comprehend the LaTeX output that will
be generated since it isn't interspersed among
numerous print statements.

Another benefit of using YAML and moving the
output to a Mustache template is that the qzm
script was reduced in size by nearly 291 lines
to just 57 lines total.

Switched from Perl to Python since Python has
easy to use libraries for YAML and Mustache.
It is also more commonly used today (IMO) which
makes this code more accessible.